refactor: move `create_document` to `extraction` submodule
