+++ server/features/extraction/helpers/__init__.py
@@ -1,0 +1,3 @@
+from server.features.extraction.helpers.create_document import (
+    create_document as create_document,
+)
